cập nhật database thuộc tính điểm của entity TraLoi

+++ API/src/Domain/Entities/TraLoi.cs
@@ -13,7 +13,7 @@ public class TraLoi
     public required string NoiDung { get; set; }
     public string? UrlFile { get; set; }
     public required string HocSinhCode { get; set; }
-    public int? Diem {  get; set; }
+    public float? Diem {  get; set; }
     public string? NhanXet {  get; set; }
     [JsonIgnore]
     public HocSinh HocSinh { get; set; } = null!;
